requirements.txt
ansible==2.5.2
appnope==0.1.0
asn1crypto==0.24.0
attrs==17.4.0
autopep8==1.3.5
backcall==0.1.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
confusable-homoglyphs==3.0.0
coverage==4.5.1
cryptography==2.2.2
decorator==4.3.0
Django==2.0.4
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.2.0
django-multiselectfield==0.1.8
django-registration==2.4.1
django-sass-processor==0.7
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.13
flake8==3.5.0
idna==2.6
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
libsass==0.14.5
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
paramiko==2.4.1
parso==0.2.0
pep8==1.7.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.3
pyasn1==0.4.2
pycodestyle==2.4.0
pycparser==2.18
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
PyNaCl==1.2.1
pytest==3.5.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
sorl-thumbnail==12.4.1
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
